+++ go/Makefile
@@ -114,7 +114,8 @@ $(foreach in,$(MOCK_STD_IN),$(eval $(call MOCK_STD_template,$(in))))
 
 bin: deps_gen
 	GOBIN=${LOCAL_GOBIN} govendor install --tags "$(GOTAGS)" -v +local,program
-	sudo setcap cap_net_admin,cap_net_raw+ep ../bin/border-acceptance
+	@sudo -p "go:braccept [sudo] password for %p: " true
+	sudo setcap cap_net_admin,cap_net_raw+ep ../bin/braccept
 
 libs: deps_gen
 	govendor install --tags "$(GOTAGS)" -v +local,^program
