--- go/border/border-acceptance/README.md
@@ -1,8 +1,0 @@
-TODO - requirements, ie. docker-compose, etc
-
-Steps to run Border Router acceptance test:
-$ cd go/border/border-acceptance
-$ ./setup.sh
-$ cd ../../..
-$ make
-$ ./bin/border-acceptance
