--- go/border/border-acceptance/br-compose.tmpl
@@ -1,14 +1,0 @@
-version: "3"
-services:
-  border:
-    image: scion_border
-    network_mode: "service:dispatcher"
-    environment:
-      SU_EXEC_USERSPEC: "$LOGNAME"
-    volumes:
-      - "/etc/passwd:/etc/passwd:ro"
-      - "/etc/group:/etc/group:ro"
-      - "/run/shm/dispatcher:/run/shm/dispatcher"
-      - "./{{ .ConfDir }}:/share/conf"
-      - "../../../logs:/share/logs"
-      command: ["-id={{ .BorderID }}", "-prom=[127.0.0.49]:30093"]
