+++ go/border/setup.go
@@ -112,11 +112,11 @@ func (r *Router) clearCapabilities() error {
 func (r *Router) loadNewConfig() (*conf.Conf, error) {
 	var config *conf.Conf
 	var err error
-	log.Debug("Loading topology config", "dir", r.confDir)
 	if config, err = conf.Load(r.Id, r.confDir); err != nil {
-		return nil, err
+		return nil, common.NewBasicError("Failed to load topology config", err, "dir", r.confDir)
 	}
-	log.Debug("Topology and AS config loaded", "IA", config.IA, "IfIDs", config.BR)
+	log.Debug("Topology and AS config loaded", "IA", config.IA, "IfIDs", config.BR,
+		"dir", r.confDir)
 	return config, nil
 }
