+++ go/lib/spkt/cmnhdr.go
@@ -53,8 +53,8 @@ func CmnHdrFromRaw(b common.RawBytes) (*CmnHdr, error) {
 func (c *CmnHdr) Parse(b common.RawBytes) error {
 	// Check for minimum header length
 	if len(b) < CmnHdrLen {
-		return common.NewBasicError("Can't parse common header, buffer is too short", nil,
-			"len", len(b))
+		return common.NewBasicError("Packet is shorter than the common header length", nil,
+			"min", CmnHdrLen, "actual", len(b))
 	}
 	offset := 0
 	verDstSrc := common.Order.Uint16(b[offset:])
